hoist uvibytes into separate module
